[refac]: renamed class "Translations" to "Translator" and renamed file "translater" to "translator"
