+++ utils/docs/translator.py
@@ -9,7 +9,7 @@ translations_file = "docs/translations/translations.json"
 translations = io.FileHandler.read(dir=SRC_DIR, file_name=translations_file)
 
 
-class Translations(BaseModel):
+class Translator(BaseModel):
     de: dict[str, str]
     fr: dict[str, str]
 
@@ -29,4 +29,4 @@ class Translations(BaseModel):
         return self
 
 
-TRANSLATE = Translations(**loads(translations))
+TRANSLATE = Translator(**loads(translations))
